+++ packages/intl/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fat-fuzzy/intl
 
+## 0.1.13
+
+### Patch Changes
+
+- 67c7f21: Improve base schemas, config AJV options, a11y fixes
+
 ## 0.1.12
 
 ### Patch Changes